requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
attrs=19.1.0=py36_1
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
bokeh=1.2.0=py36_0
ca-certificates=2019.5.15=0
certifi=2019.6.16=py36_0
cffi=1.12.3=py36h2e261b9_0
click=7.0=py36_0
cloudpickle=1.1.1=py_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cycler=0.10.0=py36_0
cytoolz=0.9.0.1=py36h14c3975_1
dask=1.2.2=py_0
dask-core=1.2.2=py_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py36_1
defusedxml=0.6.0=py_0
distributed=1.28.1=py36_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
gmp=6.1.2=hb3b607b_0
gpytorch=0.3.3=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
heapdict=1.0.0=py36_2
icu=58.2=h9c2bf20_1
intel-openmp=2019.4=243
ipykernel=5.1.1=py36h39e3cac_0
ipython=7.6.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.4.2=py36_0
jedi=0.13.3=py36_0
jinja2=2.10.1=py36_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.1=py36_0
jupyter=1.0.0=py36_7
jupyter_client=5.2.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.5.0=py_0
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
locket=0.2.0=py36_1
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.1.0=py36h5429711_0
mistune=0.8.4=py36h7b6447c_0
mkl=2018.0.3=1
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
msgpack-python=0.6.1=py36hfd86e86_1
nbconvert=5.5.0=py_0
nbformat=4.4.0=py36_0
nccl=1.3.5=cuda10.0_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
notebook=5.7.8=py36_0
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
olefile=0.46=py36_0
openssl=1.1.1=h7b6447c_0
packaging=19.0=py36_0
pandas=0.24.2=py36he6710b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.5.0=py_0
partd=0.3.10=py36_1
pcre=8.43=he6710b0_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_0
pillow=6.0.0=py36h34e0f95_0
pip=19.1.1=py36_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.9=py36_0
psutil=5.6.2=py36h7b6447c_0
ptyprocess=0.6.0=py36_0
pycparser=2.19=py36_0
pygments=2.4.2=py_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.14.11=py36h7b6447c_0
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytorch=1.1.0=py3.6_cuda10.0.130_cudnn7.5.1_0
pytz=2019.1=py_0
pyyaml=5.1=py36h7b6447c_0
pyzmq=18.0.0=py36he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.5.1=py_0
readline=7.0=h7b6447c_5
scipy=1.1.0=py36hfa4b5c9_1
send2trash=1.5.0=py36_0
setuptools=41.0.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sortedcontainers=2.1.0=py36_0
sqlite=3.28.0=h7b6447c_0
tblib=1.4.0=py_0
terminado=0.8.2=py36_0
testpath=0.4.2=py36_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py36_0
torchvision=0.3.0=py36_cu10.0.130_1
tornado=6.0.2=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
wheel=0.33.4=py36_0
widgetsnbextension=3.4.2=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zict=0.1.4=py36_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
